Rework CoreBounds on Android as iOS does
