+++ src/Uno.UI/UI/Xaml/Window.Android.cs
@@ -54,26 +54,17 @@ namespace Windows.UI.Xaml
 
 		internal void RaiseNativeSizeChanged(int screenWidth, int screenHeight)
 		{
-			int statusBarHeightDp = 0;
-
-			var activity = ContextHelper.Current as Activity;
-			var decorView = activity.Window.DecorView;
-			var isStatusBarVisible = ((int)decorView.SystemUiVisibility & (int)SystemUiFlags.Fullscreen) == 0;
-
-			if (isStatusBarVisible)
-			{
-				int resourceId = Android.Content.Res.Resources.System.GetIdentifier("status_bar_height", "dimen", "android");
-				if (resourceId > 0)
-				{
-					statusBarHeightDp = (int)(Android.Content.Res.Resources.System.GetDimensionPixelSize(resourceId) / Android.App.Application.Context.Resources.DisplayMetrics.Density);
-				}
-			}
-
 			var newBounds = ViewHelper.PhysicalToLogicalPixels(new Rect(0, 0, screenWidth, screenHeight));
 
 			if (Bounds != newBounds)
 			{
 				Bounds = newBounds;
+
+				var applicationView = ApplicationView.GetForCurrentView();
+				if (applicationView != null)
+				{
+					applicationView.SetCoreBounds(newBounds);
+				}
 
 				RaiseSizeChanged(
 					new WindowSizeChangedEventArgs(
@@ -81,10 +72,6 @@ namespace Windows.UI.Xaml
 					)
 				);
 			}
-
-			var visibleBoundsConfig = Android.App.Application.Context.Resources.Configuration;
-			var newVisibleBounds = new Rect(0, statusBarHeightDp, visibleBoundsConfig.ScreenWidthDp, visibleBoundsConfig.ScreenHeightDp);
-			ApplicationView.GetForCurrentView().SetVisibleBounds(newVisibleBounds);
 		}
 	}
 }
